[x86_64 boot] Use `usize` for addresses instead of `u64`
These two types will always be exactly the same on x86_64 anyways, but
apparently `usize` is specifically meant for representing pointer values,
so it seems more fitting for addresses to be `usize`. This makes it a lot
easier to use with libcore memory stuff like core::alloc

+++ bootloader/src/boot_services/memory.rs
@@ -18,7 +18,7 @@ impl BootServices {
     pub fn allocate_frames(
         &self,
         memory_type: MemoryType,
-        pages: u64,
+        pages: usize,
     ) -> Result<PhysicalAddress, Status> {
         let mut start_address = PhysicalAddress::default();
         match (self._allocate_pages)(
@@ -35,7 +35,7 @@ impl BootServices {
     }
 
     /// Frees memory pages
-    pub fn free_pages(&self, memory: PhysicalAddress, pages: u64) -> Result<(), Status> {
+    pub fn free_pages(&self, memory: PhysicalAddress, pages: usize) -> Result<(), Status> {
         (self._free_pages)(memory, pages).as_result().map(|_| ())
     }
